+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 ## [Unreleased]
 
+### Features
+
+- Can now read and convert from `STDIN`
+
+### Breaking
+
+- `--std-out` is now `--stdout`
+
+### Internal
+
+- Switched from structopt to stdin
+- Upgraded dependencies
+
 ## [v0.1.2] - 2021-09-29
+
 - Upgraded dependencies
 
 ## [v0.1.1] - 2021-09-29